+++ p2panda-rs/src/document/document_view_id.rs
@@ -35,15 +35,33 @@ pub struct DocumentViewId(Vec<OperationId>);
 
 impl DocumentViewId {
     /// Create a new document view id.
-    pub fn new(graph_tips: Vec<OperationId>) -> Self {
-        let mut graph_tips_mut = graph_tips;
-        graph_tips_mut.sort();
-        Self(graph_tips_mut)
+    pub fn new(graph_tips: &[OperationId]) -> Self {
+        Self(graph_tips.to_vec())
     }
 
     /// Get the graph tip ids of this view id.
     pub fn graph_tips(&self) -> &[OperationId] {
         self.0.as_slice()
+    }
+
+    /// Returns a hash over the sorted graph tips constituting this view id.
+    ///
+    /// Use this as a unique identifier for a document if you need a value with a limited size. The
+    /// document view id itself grows with the number of graph tips that the document has, which
+    /// may not be desirable for an identifier.
+    ///
+    /// Keep in mind that when you refer to document views with this hash value it will not be
+    /// possible to recover the document view id from it.
+    pub fn hash(&self) -> Hash {
+        // Sort graph tips to ensure consistent hashes
+        let mut graph_tips_mut = self.0.clone();
+        graph_tips_mut.sort();
+
+        let graph_tip_bytes = graph_tips_mut
+            .into_iter()
+            .flat_map(|graph_tip| graph_tip.as_hash().to_bytes())
+            .collect();
+        Hash::new_from_bytes(graph_tip_bytes).unwrap()
     }
 }
 
@@ -84,7 +102,7 @@ impl IntoIterator for DocumentViewId {
 /// only consists of one graph tip hash.
 impl From<OperationId> for DocumentViewId {
     fn from(operation_id: OperationId) -> Self {
-        Self::new(vec![operation_id])
+        Self::new(&[operation_id])
     }
 }
 
@@ -94,7 +112,7 @@ impl From<OperationId> for DocumentViewId {
 /// consists of one graph tip hash.
 impl From<Hash> for DocumentViewId {
     fn from(hash: Hash) -> Self {
-        Self::new(vec![hash.into()])
+        Self::new(&[hash.into()])
     }
 }
 
@@ -106,7 +124,7 @@ impl FromStr for DocumentViewId {
     type Err = DocumentViewIdError;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Ok(Self::new(vec![Hash::new(s)?.into()]))
+        Ok(Self::new(&[Hash::new(s)?.into()]))
     }
 }
 
@@ -128,16 +146,16 @@ mod tests {
         let document_id: DocumentViewId = hash_str.parse().unwrap();
         assert_eq!(
             document_id,
-            DocumentViewId::new(vec![hash_str.parse::<OperationId>().unwrap()])
+            DocumentViewId::new(&[hash_str.parse::<OperationId>().unwrap()])
         );
 
         // Converts a `Hash` to `DocumentViewId`
         let document_id: DocumentViewId = hash.clone().into();
-        assert_eq!(document_id, DocumentViewId::new(vec![hash.clone().into()]));
+        assert_eq!(document_id, DocumentViewId::new(&[hash.clone().into()]));
 
         // Converts an `OperationId` to `DocumentViewId`
         let document_id: DocumentViewId = OperationId::new(hash.clone()).into();
-        assert_eq!(document_id, DocumentViewId::new(vec![hash.into()]));
+        assert_eq!(document_id, DocumentViewId::new(&[hash.into()]));
 
         // Fails when string is not a hash
         assert!("This is not a hash".parse::<DocumentViewId>().is_err());
@@ -155,8 +173,8 @@ mod tests {
         #[from(random_operation_id)] operation_id_1: OperationId,
         #[from(random_operation_id)] operation_id_2: OperationId,
     ) {
-        let view_id_1 = DocumentViewId::new(vec![operation_id_1.clone(), operation_id_2.clone()]);
-        let view_id_2 = DocumentViewId::new(vec![operation_id_2, operation_id_1]);
+        let view_id_1 = DocumentViewId::new(&[operation_id_1.clone(), operation_id_2.clone()]);
+        let view_id_2 = DocumentViewId::new(&[operation_id_2, operation_id_1]);
         assert_eq!(view_id_1, view_id_2);
     }
 
@@ -178,5 +196,17 @@ mod tests {
             format!("{}", view_id_1.validate().unwrap_err()),
             "Expected sorted operation ids in document view id"
         );
+    }
+
+    #[rstest]
+    fn document_view_hash(
+        #[from(random_operation_id)] operation_id_1: OperationId,
+        #[from(random_operation_id)] operation_id_2: OperationId,
+    ) {
+        let view_id_1 = DocumentViewId::new(&[operation_id_1.clone(), operation_id_2.clone()]);
+        assert!(view_id_1.hash().validate().is_ok());
+
+        let view_id_2 = DocumentViewId::new(&[operation_id_2, operation_id_1]);
+        assert_eq!(view_id_1.hash(), view_id_2.hash());
     }
 }